# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "project akhir",
    "description": "",
    "contact": {}
  },
  "host": "localhost:8000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/seller/product": {
      "post": {
        "description": "",
        "summary": "Add Product",
        "tags": [
          "seller"
        ],
        "operationId": "AddProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "nama",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "harga",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "keterangan",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "id_kategori",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "gambar",
            "in": "formData",
            "required": true,
            "type": "file",
            "format": "file",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "description": "",
        "summary": "Get All Product",
        "tags": [
          "seller"
        ],
        "operationId": "GetAllProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/v1/seller/order/1": {
      "put": {
        "description": "",
        "summary": "Update Status Order",
        "tags": [
          "seller"
        ],
        "operationId": "UpdateStatusOrder",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateStatusOrderRequest"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/v1/buyer/order/1": {
      "post": {
        "description": "",
        "summary": "Add Order",
        "tags": [
          "buyer"
        ],
        "operationId": "AddOrder",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddOrderRequest"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/buyer/wishlist": {
      "post": {
        "description": "",
        "summary": "Add Wishlist",
        "tags": [
          "buyer"
        ],
        "operationId": "AddWishlist",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddWishlistRequest"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "description": "",
        "summary": "Get All Wishlist",
        "tags": [
          "buyer"
        ],
        "operationId": "GetAllWishlist",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/buyer/wishlist/53": {
      "delete": {
        "description": "",
        "summary": "Delete Wishlist",
        "tags": [
          "buyer"
        ],
        "operationId": "DeleteWishlist",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "default": ""
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "definitions": {
...
    "UpdateStatusOrderRequest": {
      "title": "UpdateStatusOrderRequest",
      "example": {
        "status": true
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean"
        }
      },
      "required": [
        "status"
      ]
    },
...
    "AddOrderRequest": {
      "title": "AddOrderRequest",
      "example": {
        "harga_tawar": 2000
      },
      "type": "object",
      "properties": {
        "harga_tawar": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "harga_tawar"
      ]
    },
    "AddWishlistRequest": {
      "title": "AddWishlistRequest",
      "example": {
        "idProduct": 2
      },
      "type": "object",
      "properties": {
        "idProduct": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "idProduct"
      ]
    }
  },
...
}